+++ src/Cartalyst/Sentry/Users/Exceptions.php
@@ -24,3 +24,4 @@ class UserAlreadyActivatedException extends \RuntimeException {}
 class UserNotFoundException extends \OutOfBoundsException {}
 class UserNotActivatedException extends \RuntimeException {}
 class UserExistsException extends \UnexpectedValueException {}
+class WrongPasswordException extends UserNotFoundException {}